move ensure_comspec_set to _nsis.py because it must stand alone
